backend/requirements.txt
langgraph
langgraph-cli
langchain-google-genai
langchain-mcp-adapters
mcp
python-dotenv
uvicorn
